Move get_object_if_allowed to core.guardian
